refactor: app에서 router 분리
