+++ src/App.tsx
@@ -1,18 +1,9 @@
-import { RouterProvider, createBrowserRouter } from 'react-router-dom';
-import { HomePage } from './pages';
-import { PATH } from './constants';
+import { RouterProvider } from 'react-router-dom';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import { router } from './routes';
 
 const queryClient: QueryClient = new QueryClient();
-
-const router = createBrowserRouter([
-  {
-    path: PATH.HOME,
-    element: <HomePage />,
-    errorElement: <div>Error Page</div>
-  }
-]);
 
 const App = () => {
   return (
