+++ src/constants/index.ts
@@ -1,1 +1,1 @@
-export * from './path';
+export * from '../routes/constants';
